pyproject.toml
[tool.poetry]
name = "ecommerce_backend"
version = "0.1.0"
description = ""
authors = ["avinash yadav <avinash.yadav@techvzero.com>"]

[tool.poetry.dependencies]
python = "^3.6"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
pydantic = "^1.9.1"
SQLAlchemy = "^1.3.15"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^3.2.2"
python-multipart = "^0.0.5"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "runserver:main"
